apps/orchestrator-api/requirements.txt
# PolPilot Orchestrator API — Requirements
anthropic>=0.52.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
httpx>=0.28.0
python-dotenv>=1.1.0
